Metal 2.0: split ComputeConfiguration and DataMovementConfiguration into own headers
kernel_spec.hpp previously opened with ~75 lines of ComputeConfiguration and
DataMovementConfiguration definitions before reaching the KernelSpec orientation
doc and struct. That violated general-to-specific reading order: readers had to
scroll past hairy pipeline / DM hardware config details to find out what the
file was actually about.

Split:
  - compute_configuration.hpp       — ComputeConfiguration
  - data_movement_configuration.hpp — DataMovementConfiguration + Gen1/Gen2 glossary
  - kernel_spec.hpp                 — KernelSpec + aliases + DFB factories

Matches the existing pattern in the directory (dataflow_buffer_spec.hpp,
semaphore_spec.hpp, tensor_parameter.hpp, node_coord.hpp). Users still
#include <.../kernel_spec.hpp> and transitively get everything — no
include changes in impl or tests.

Forward declarations would have been an alternative but don't work for
std::variant<ComputeConfiguration, DataMovementConfiguration>: variant
member instantiation requires both alternatives complete.

Full repo builds clean.
